refactor: 書き方をAdmin Gateway側に揃える
