+++ api/gateway/native/src/api/auth.ts
