+++ api/gateway/native/src/api/index.ts
@@ -1,1 +1,1 @@
-export * from './user'
+export * from './auth'
