+++ api/gateway/native/src/types/input/auth.ts
