+++ api/gateway/native/src/types/input/index.ts
@@ -1,1 +1,1 @@
-export * from './user'
+export * from './auth'
