+++ api/gateway/native/src/types/output/auth.ts
