+++ api/gateway/native/src/types/output/index.ts
@@ -1,1 +1,1 @@
-export * from './user'
+export * from './auth'
